refact: move delimiters to their respective styles
