+++ core/src/test/java/com/jeanbarrossilva/orca/core/feed/profile/toot/style/styling/bold/test/ColonBoldDelimiter.kt
@@ -1,8 +1,8 @@
 package com.jeanbarrossilva.orca.core.feed.profile.toot.style.styling.bold.test
 
-import com.jeanbarrossilva.orca.core.feed.profile.toot.style.styling.bold.BoldDelimiter
+import com.jeanbarrossilva.orca.core.feed.profile.toot.style.styling.bold.Bold
 
-internal object ColonBoldDelimiter : BoldDelimiter() {
+internal object ColonBoldDelimiter : Bold.Delimiter.Child() {
     override val regex = Regex(":(.*?):")
 
     override fun onGetTarget(match: String): String {
