+++ core/src/test/java/com/jeanbarrossilva/orca/core/feed/profile/toot/style/styling/mention/test/ColonMentionDelimiter.kt
@@ -1,8 +1,8 @@
 package com.jeanbarrossilva.orca.core.feed.profile.toot.style.styling.mention.test
 
-import com.jeanbarrossilva.orca.core.feed.profile.toot.style.styling.mention.MentionDelimiter
+import com.jeanbarrossilva.orca.core.feed.profile.toot.style.styling.mention.Mention
 
-internal object ColonMentionDelimiter : MentionDelimiter() {
+internal object ColonMentionDelimiter : Mention.Delimiter.Child() {
     override val regex = Regex(":[a-zA-Z0-9._%+-]+")
 
     override fun onGetTarget(match: String): String {
